Fix compilation for modern Linux (GCC 13, C++14)
This commit fixes numerous compatibility issues to allow StepMania 3.9
to compile on modern Linux systems with GCC 13 and glibc 2.39.

Build System Changes:
- autoconf/m4/lua.m4: Detect Lua 5.1 via pkg-config (modern naming)
- autoconf/m4/opengl.m4: Always include -lX11 in standard paths
- src/Makefile.am: Add -std=c++14 and -fpermissive flags

Header Fixes:
- src/global.h: Add #include <cstring> for memset/strcmp
- src/RageUtil.h: Add #include <climits> for INT_MAX
- src/arch/ArchHooks/ArchHooks.h: Add #include <ctime> for tm struct
- src/LuaFunctions.h: Add #include <lauxlib.h> for Lua 5.1 compatibility

C++ Compatibility Fixes:
- Remove extra class qualifications (GameState::, FilenameDB::, etc.)
- Fix void* to int casts using intptr_t for 64-bit systems
- Fix bool to pointer implicit conversions (return NULL instead of false)
- Add this-> prefix for template base class member lookups
- Rename error_t struct to avoid glibc conflict

Library API Updates:
- libpng: Use png_get_io_ptr/png_get_error_ptr accessor functions
- libpng: Use png_set_expand_gray_1_2_4_to_8 (renamed function)
- SDL: Fix SDL_EventMask variable name conflict with SDL enum
- crypto++: Move byte typedef into CryptoPP namespace (std::byte conflict)

Platform Fixes:
- LinuxThreadHelpers.cpp: Replace _syscall0 with syscall(SYS_gettid)
- LowLevelWindow_SDL.cpp: Check both 'unix' and 'UNIX' defines
- OpenGL: Use system GL headers on Linux, remove bundled glext.h conflicts

Documentation:
- Add CLAUDE.md with build instructions and architecture overview

+++ src/arch/Threads/Threads_Pthreads.cpp
@@ -44,7 +44,7 @@ int ThreadImpl_Pthreads::Wait()
 	if( ret )
 		RageException::Throw( "pthread_join: %s", strerror(errno) );
 
-	return (int) val;
+	return (int)(intptr_t) val;
 }
 
 ThreadImpl *MakeThisThread()
